Small Refactor, Nothing Really Changed

+++ Abstract/src/Abstract/math/Matrix.cpp
@@ -46,13 +46,13 @@ namespace Abstract {
 		return (*this);
 	}
 
-	Matrix4& Matrix4::setPerspective(float fov, float ar, float near, float far)
+	Matrix4& Matrix4::setPerspective(double fov, double ar, double near, double far)
 	{
-		float tanHalfFov = tanf(fov / 2.0f);
-		float zRange = near - far;
+		double tanHalfFov = tanf(fov / 2.0f);
+		double zRange = near - far;
 
-		m[0][0] = 1.0f / (tanHalfFov * ar); m[1][0] = 0;				 m[2][0] = 0;					   m[3][0] = 0;
-		m[0][1] = 0;						m[1][1] = 1.0f / tanHalfFov; m[2][1] = 0;					   m[3][1] = 0;
+		m[0][0] = 1.0 / (tanHalfFov * ar);  m[1][0] = 0;				 m[2][0] = 0;					   m[3][0] = 0;
+		m[0][1] = 0;						m[1][1] = 1.0 / tanHalfFov;  m[2][1] = 0;					   m[3][1] = 0;
 		m[0][2] = 0;						m[1][2] = 0;				 m[2][2] = (-near - far) / zRange; m[3][2] = 2 * far * near / zRange;
 		m[0][3] = 0;						m[1][3] = 0;				 m[2][3] = 1;					   m[3][3] = 0;
