+++ Abstract/src/Abstract/math/Vector.h
@@ -7,31 +7,31 @@ namespace Abstract {
 	class ABSTRACT_API Vector3
 	{
 	public:
-		Vector3(float x = 0.0f, float y = 0.0f, float z = 0.0f) : x(x), y(y), z(z) { }
+		Vector3(double x = 0.0f, double y = 0.0f, double z = 0.0f) : x(x), y(y), z(z) { }
 
-		constexpr float getX() const { return x; }
-		constexpr float getY() const { return y; }
-		constexpr float getZ() const { return z; }
+		constexpr double getX() const { return x; }
+		constexpr double getY() const { return y; }
+		constexpr double getZ() const { return z; }
 
-		constexpr void setX(float x) { this->x = x; }
-		constexpr void setY(float y) { this->y = y; }
-		constexpr void setZ(float z) { this->z = z; }
+		constexpr void setX(double x) { this->x = x; }
+		constexpr void setY(double y) { this->y = y; }
+		constexpr void setZ(double z) { this->z = z; }
 
-		inline float length() const { return sqrtf(x * x + y * y + z * z); }
+		inline double length() const { return sqrt(x * x + y * y + z * z); }
 		inline Vector3 normalized() const { return this->operator/(length()); }
 
-		constexpr float dot(const Vector3& other) const { return x * other.x + y * other.y + z * other.z; }
+		constexpr double dot(const Vector3& other) const { return x * other.x + y * other.y + z * other.z; }
 		inline Vector3 cross(const Vector3& other) const
 		{
-			float _x = y * other.z - z * other.y;
-			float _y = z * other.x - x * other.z;
-			float _z = x * other.y - y * other.x;
+			double _x = y * other.z - z * other.y;
+			double _y = z * other.x - x * other.z;
+			double _z = x * other.y - y * other.x;
 
 			return Vector3(_x, _y, _z);
 		}
 
-		inline float min() const { return fminf(x, fminf(y, z)); }
-		inline float max() const { return fmaxf(x, fmaxf(y, z)); }
+		inline double min() const { return fmin(x, fmin(y, z)); }
+		inline double max() const { return fmax(x, fmax(y, z)); }
 
 #pragma region Operator Overloads
 
@@ -40,7 +40,7 @@ namespace Abstract {
 			return Vector3(-x, -y, -z);
 		}
 
-		inline Vector3 operator+(const float& other) const
+		inline Vector3 operator+(const double& other) const
 		{
 			return Vector3(x + other, y + other, z + other);
 		}
@@ -50,7 +50,7 @@ namespace Abstract {
 			return Vector3(x + other.x, y + other.y, z + other.z);
 		}
 
-		inline Vector3 operator-(const float& other) const
+		inline Vector3 operator-(const double& other) const
 		{
 			return Vector3(x - other, y - other, z - other);
 		}
@@ -60,7 +60,7 @@ namespace Abstract {
 			return Vector3(x - other.x, y - other.y, z - other.z);
 		}
 
-		inline Vector3 operator*(const float& other) const
+		inline Vector3 operator*(const double& other) const
 		{
 			return Vector3(x * other, y * other, z * other);
 		}
@@ -70,7 +70,7 @@ namespace Abstract {
 			return Vector3(x * other.x, y * other.y, z * other.z);
 		}
 
-		inline Vector3 operator/(const float& other) const
+		inline Vector3 operator/(const double& other) const
 		{
 			return Vector3(x / other, y / other, z / other);
 		}
@@ -80,7 +80,7 @@ namespace Abstract {
 			return Vector3(x / other.x, y / other.y, z / other.z);
 		}
 
-		constexpr void operator+=(const float& other)
+		constexpr void operator+=(const double& other)
 		{
 			this->x += other;
 			this->y += other;
@@ -94,7 +94,7 @@ namespace Abstract {
 			this->z += other.z;
 		}
 
-		constexpr void operator-=(const float& other)
+		constexpr void operator-=(const double& other)
 		{
 			this->x -= other;
 			this->y -= other;
@@ -108,7 +108,7 @@ namespace Abstract {
 			this->z -= other.z;
 		}
 
-		constexpr void operator*=(const float& other)
+		constexpr void operator*=(const double& other)
 		{
 			this->x *= other;
 			this->y *= other;
@@ -122,7 +122,7 @@ namespace Abstract {
 			this->z *= other.z;
 		}
 
-		constexpr void operator/=(const float& other)
+		constexpr void operator/=(const double& other)
 		{
 			this->x /= other;
 			this->y /= other;
@@ -139,34 +139,34 @@ namespace Abstract {
 #pragma endregion
 		~Vector3() { }
 	private:
-		float x;
-		float y;
-		float z;
+		double x;
+		double y;
+		double z;
 	};
 
 	class ABSTRACT_API Vector2
 	{
 	public:
-		Vector2(float x = 0.0f, float y = 0.0f) : x(x), y(y) { }
+		Vector2(double x = 0.0f, double y = 0.0f) : x(x), y(y) { }
 
-		constexpr float getX() const { return x; }
-		constexpr float getY() const { return y; }
+		constexpr double getX() const { return x; }
+		constexpr double getY() const { return y; }
 
-		constexpr void setX(float x) { this->x = x; }
-		constexpr void setY(float y) { this->y = y; }
+		constexpr void setX(double x) { this->x = x; }
+		constexpr void setY(double y) { this->y = y; }
 
-		inline float length() const { return sqrtf(x * x + y * y); }
+		inline double length() const { return sqrt(x * x + y * y); }
 		inline Vector2 normalized() const { return this->operator/(length()); }
 
-		constexpr float dot(const Vector2& other) const { return x * other.x + y * other.y; }
-		constexpr float cross(const Vector2& other) const { return x * other.y - y * other.x; }
+		constexpr double dot(const Vector2& other) const { return x * other.x + y * other.y; }
+		constexpr double cross(const Vector2& other) const { return x * other.y - y * other.x; }
 
-		inline float min() { return fminf(x, y); }
-		inline float max() { return fmaxf(x, y); }
+		inline double min() { return fmin(x, y); }
+		inline double max() { return fmax(x, y); }
 
 #pragma region Operator Overloads
 
-		inline Vector2 operator+(const float& other) const
+		inline Vector2 operator+(const double& other) const
 		{
 			return Vector2(x + other, y + other);
 		}
@@ -176,7 +176,7 @@ namespace Abstract {
 			return Vector2(x + other.x, y + other.y);
 		}
 
-		inline Vector2 operator-(const float& other) const
+		inline Vector2 operator-(const double& other) const
 		{
 			return Vector2(x - other, y - other);
 		}
@@ -186,7 +186,7 @@ namespace Abstract {
 			return Vector2(x - other.x, y - other.y);
 		}
 
-		inline Vector2 operator*(const float& other) const
+		inline Vector2 operator*(const double& other) const
 		{
 			return Vector2(x * other, y * other);
 		}
@@ -196,7 +196,7 @@ namespace Abstract {
 			return Vector2(x * other.x, y * other.y);
 		}
 
-		inline Vector2 operator/(const float& other) const
+		inline Vector2 operator/(const double& other) const
 		{
 			return Vector2(x / other, y / other);
 		}
@@ -206,7 +206,7 @@ namespace Abstract {
 			return Vector2(x / other.x, y / other.y);
 		}
 
-		constexpr void operator+=(const float& other)
+		constexpr void operator+=(const double& other)
 		{
 			this->x += other;
 			this->y += other;
@@ -218,7 +218,7 @@ namespace Abstract {
 			this->y += other.y;
 		}
 
-		constexpr void operator-=(const float& other)
+		constexpr void operator-=(const double& other)
 		{
 			this->x -= other;
 			this->y -= other;
@@ -230,7 +230,7 @@ namespace Abstract {
 			this->y -= other.y;
 		}
 
-		constexpr void operator*=(const float& other)
+		constexpr void operator*=(const double& other)
 		{
 			this->x *= other;
 			this->y *= other;
@@ -242,7 +242,7 @@ namespace Abstract {
 			this->y *= other.y;
 		}
 
-		constexpr void operator/=(const float& other)
+		constexpr void operator/=(const double& other)
 		{
 			this->x /= other;
 			this->y /= other;
@@ -257,8 +257,8 @@ namespace Abstract {
 #pragma endregion
 		~Vector2() { }
 	private:
-		float x;
-		float y;
+		double x;
+		double y;
 	};
 
 }
